+++ src/server/devices/baseDeviceInterfaces/iHeater.ts
@@ -1,18 +1,18 @@
-import { HeaterSettings, TemperaturSettings } from '../../../models';
+import { HeaterSettings, TemperatureSettings } from '../../../models';
 import { IBaseDevice } from './iBaseDevice';
 
 export interface iHeater extends IBaseDevice {
   settings: HeaterSettings;
-  readonly desiredTemperatur: number;
+  readonly desiredTemperature: number;
   readonly humidity: number;
   readonly iLevel: number;
-  readonly iTemperatur: number;
+  readonly iTemperature: number;
 
   set seasonTurnOff(value: boolean);
 
   deleteAutomaticPoint(name: string): void;
 
-  setAutomaticPoint(name: string, setting: TemperaturSettings): void;
+  setAutomaticPoint(name: string, setting: TemperatureSettings): void;
 
   stopAutomaticCheck(): void;